requirements.txt
Django==1.7
pillow==2.7.0
wheel==0.24.0
django-registration-redux==1.1
